feat: added extensions, lintd
